shev-session-ser/requirements.txt
bleach==2.1.2
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==1.7.2
decorator==4.2.1
Django==1.11.9
django-redis==4.8.0
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
entrypoints==0.2.3
ez==2.8.7
ez-setup==0.9
flake8==3.5.0
html5lib==1.0.1
idna==2.6
ipykernel==4.8.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jwt==0.5.2
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
mysql-connector==2.1.4
mysql-connector-python==8.0.6
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
pandocfilters==1.4.2
parso==0.1.1
pew==1.1.2
pexpect==4.3.1
pickleshare==0.7.4
pipenv==9.0.3
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.5.3
PyMySQL==0.8.0
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.4
qtconsole==4.3.1
redis==2.10.6
requests==2.18.4
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.3
